[loaders] Split API code into multiple files and duplicate the legacy `http`  code into the new module.
