+++ lib/loaders/loaders_with_authentication.js
@@ -1,4 +1,4 @@
-import { gravityLoaderWithAuthenticationFactory, impulseLoaderWithAuthenticationFactory } from "./apis"
+import { gravityLoaderWithAuthenticationFactory, impulseLoaderWithAuthenticationFactory } from "./api"
 const { IMPULSE_APPLICATION_ID } = process.env
 
 export default (accessToken, userID) => {
