+++ lib/loaders/loaders_without_authentication.js
@@ -1,7 +1,7 @@
 import {
   gravityLoaderWithoutAuthenticationFactory as gravityLoader,
   positronLoaderWithoutAuthenticationFactory as positronLoader,
-} from "./apis"
+} from "./api"
 
 export default () => ({
   articlesLoader: positronLoader("articles"),
