+++ test/lib/loaders/api.js
@@ -1,4 +1,6 @@
-import { apiLoaderWithAuthenticationFactory, apiLoaderWithoutAuthenticationFactory } from "lib/loaders/apis"
+import { apiLoaderWithAuthenticationFactory } from "lib/loaders/api/loader_with_authentication_factory"
+import { apiLoaderWithoutAuthenticationFactory } from "lib/loaders/api/loader_without_authentication_factory"
+
 import cache from "lib/cache"
 
 describe("API loaders", () => {
